Allow setting SSL verification at the class level
Since some wiski/kiwis servers have certificate issues we want to be
able to either disable SSL verification or provide an SSL certificate
that works for the given server. Since this is a problem at the server
level moving the passing of the flag from the `KIWIS` class methods to
the `KIWIS` class init method. This allows for setting the SSL
verification once at object creation instead of for each individual
call.